+++ cl_input.c
@@ -37,6 +37,7 @@ cvar_t cl_anglespeedkey       = {"cl_anglespeedkey","1.5"};
 cvar_t cl_backspeed           = {"cl_backspeed","400"};
 cvar_t cl_c2spps              = {"cl_c2spps","0"};
 cvar_t cl_c2sImpulseBackup    = {"cl_c2sImpulseBackup","3"};
+cvar_t cl_c2sdupe             = {"cl_c2sdupe", "0"};
 cvar_t cl_forwardspeed        = {"cl_forwardspeed","400"};
 cvar_t cl_smartjump           = {"cl_smartjump", "1"};
 cvar_t cl_iDrive              = {"cl_iDrive", "0", 0, Rulesets_OnChange_cl_iDrive};
@@ -1307,6 +1308,9 @@ void CL_SendCmd(void)
 	// network stats table
 	network_stats[cls.netchan.outgoing_sequence&NETWORK_STATS_MASK].sentsize = buf.cursize + 8;
 
+	//send duplicated packets, if set
+	cls.netchan.dupe = bound(0, cl_c2sdupe.value, MAX_DUPLICATE_PACKETS);
+
 	// deliver the message
 	Netchan_Transmit (&cls.netchan, buf.cursize, buf.data);
 }
@@ -1392,11 +1396,11 @@ void CL_InitInput(void)
 	Cvar_Register(&m_accel_offset);
 	Cvar_Register(&m_accel_senscap);
 
-
 	Cvar_SetCurrentGroup(CVAR_GROUP_NETWORK);
 	Cvar_Register(&cl_nodelta);
 	Cvar_Register(&cl_c2sImpulseBackup);
 	Cvar_Register(&cl_c2spps);
+	Cvar_Register(&cl_c2sdupe);
 
 	Cvar_ResetCurrentGroup();